+++ package.json
@@ -12,6 +12,7 @@
     "@notionhq/client": "^2.2.14",
     "next": "latest",
     "next-auth": "^4.24.7",
+    "node": "^18.20.0",
     "react": "latest",
     "react-dom": "latest",
     "react-icons": "^5.0.1",
