+++ src/components/Mission_Statement/index.tsx
